Turn basic components into a standalone widget for future reuse
