+++ tomviz/acquisition/PassiveAcquisitionWidget.h
@@ -18,6 +18,8 @@
 
 #include <QDialog>
 
+#include "MatchInfo.h"
+
 #include <QLabel>
 #include <QMap>
 #include <QPointer>
@@ -45,15 +47,6 @@ namespace tomviz {
 class AcquisitionClient;
 class DataSource;
 
-enum class TestRegexFormat {
-  npDm3,
-  pmDm3,
-  npTiff,
-  pmTiff,
-  Custom,
-  Advanced
-};
-
 class PassiveAcquisitionWidget : public QDialog
 {
   Q_OBJECT
@@ -76,10 +69,10 @@ private slots:
   void onError(const QString& errorMessage, const QJsonValue& errorData);
   void watchSource();
 
-  void formatChanged(int);
   void formatTabChanged(int index);
   void testFileNameChanged(QString);
-  void customFileRegex();
+
+  void onRegexChanged(QString);
 
 signals:
   void connectParameterDescription(QJsonValue params);
@@ -88,11 +81,7 @@ private:
   QScopedPointer<Ui::PassiveAcquisitionWidget> m_ui;
   QScopedPointer<AcquisitionClient> m_client;
 
-  QRegExp m_fileNameRegex;
   QString m_testFileName;
-  QString m_negChar;
-  QString m_posChar;
-  QString m_pythonFileNameRegex;
 
   vtkNew<vtkRenderer> m_renderer;
   vtkNew<vtkInteractorStyleRubberBand2D> m_defaultInteractorStyle;
@@ -119,22 +108,10 @@ private:
   void startLocalServer();
   void displayError(const QString& errorMessage);
   void stopWatching();
+  void validateTestFileName();
 
-  void validateFileNameRegex();
   void setupTestTable();
   void resizeTestTable();
-  void buildFileRegex(QString, QString, QString, QString, QString);
-  void setupFileFormatCombo();
-
-  QList<TestRegexFormat> makeDefaultFormatOrder() const;
-  QMap<TestRegexFormat, QString> makeDefaultFileNames() const;
-  QMap<TestRegexFormat, QString> makeDefaultLabels() const;
-  QMap<TestRegexFormat, QStringList> makeDefaultRegexParams() const;
-
-  const QList<TestRegexFormat> m_defaultFormatOrder;
-  const QMap<TestRegexFormat, QString> m_defaultFileNames;
-  const QMap<TestRegexFormat, QString> m_defaultFormatLabels;
-  const QMap<TestRegexFormat, QStringList> m_defaultRegexParams;
 };
 } // namespace tomviz
