+++ tomviz/acquisition/PassiveAcquisitionWidget.ui
@@ -51,51 +51,10 @@
    </item>
    <item row="3" column="0" colspan="2">
     <widget class="QTabWidget" name="formatTabWidget">
-     <widget class="QWidget" name="basicTab">
+     <widget class="tomviz::BasicFormatWidget" name="basicTab" native="true">
       <attribute name="title">
        <string>Basic</string>
       </attribute>
-      <layout class="QGridLayout" name="basicTabLayout">
-       <item row="0" column="0">
-        <widget class="QLabel" name="fileNameFormatLabel">
-         <property name="text">
-          <string>File Name Format</string>
-         </property>
-        </widget>
-       </item>
-       <item row="0" column="1">
-        <widget class="QComboBox" name="fileFormatCombo">
-         <property name="enabled">
-          <bool>true</bool>
-         </property>
-        </widget>
-       </item>
-       <item row="1" column="0" colspan="2">
-        <widget class="QGroupBox" name="customGroupBox">
-         <property name="title">
-          <string>Custom format</string>
-         </property>
-         <layout class="QVBoxLayout" name="verticalLayout_group">
-          <item>
-           <widget class="tomviz::CustomFormatWidget" name="customFormatWidget" native="true"/>
-          </item>
-         </layout>
-        </widget>
-       </item>
-       <item row="2" column="0" colspan="2">
-        <spacer name="vBottomSpacer3">
-         <property name="orientation">
-          <enum>Qt::Vertical</enum>
-         </property>
-         <property name="sizeHint" stdset="0">
-          <size>
-           <width>0</width>
-           <height>30</height>
-          </size>
-         </property>
-        </spacer>
-       </item>
-      </layout>
      </widget>
      <widget class="tomviz::AdvancedFormatWidget" name="advancedTab" native="true">
       <attribute name="title">
@@ -197,9 +156,9 @@
    <container>1</container>
   </customwidget>
   <customwidget>
-   <class>tomviz::CustomFormatWidget</class>
+   <class>tomviz::BasicFormatWidget</class>
    <extends>QWidget</extends>
-   <header>CustomFormatWidget.h</header>
+   <header>BasicFormatWidget.h</header>
    <container>1</container>
   </customwidget>
   <customwidget>
